Drop support for node version older than 8
In node 8 `lib/_http_outgoing.js` was changed in node core. The current
implementation works against node 8 and above, but I have not
been able to diagnose exactly what breaks in the old version.

It seems to be something with res.write being bound in the wrong way -
at least, what happens is that the overwritten version of res.write is
called from within the original res.end.

Since node 6 has been end of lifed for almost a couple of years, I will
not spend further time debugging it and just bump the minimal required
node version... and schedule a code-style overhaul, since we are now
officially not supporting any non ES6-capable node versions.