Add destoryAndRestore method to reponseBody stream
In express-processimage res.unhijack() was used as a way to discard the
stream and allow normal error handling by passing errors to `next` to
work again.

That was an unsafe/unsupported operation in the old implementation, and
it would at the very least result in dangling half-read streams.

This implementation, with this method added, should address any issues
with that usage pattern.