Proxy statusCode and statusMessage to res
This is not happening as we are using an empty object with res as it's
proxy type. It would be really cool to be able to use a Proxy here, but
I will have to spend a little more time getting that right.

A proxy will be much more stable than this sort of hackery.